chore: add additional test cases
